backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
python-decouple>=3.8
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<22.0
django-filter>=23.0,<24.0
drf-yasg>=1.21,<2.0
